+++ greenfield/eldenring/tables/item_ids.py
@@ -12003,4 +12003,4 @@ JUNK_GEAR_NAMES = [
     'Zamor Mask',
 ]
 
-_GEN_STAMP = {'inputs_hash': 'sha256:e37ef453818b940e0e1e1460d4745c231839819a9f32fc80ef77a91bc6b48622', 'module': 'item_ids.py', 'body_sha256': 'sha256:2158491c6a50bfdca082dff40637c578a45e31fbf1c44c3d3847f8c88e747de3'}
+_GEN_STAMP = {'inputs_hash': 'sha256:f1402adc277de2ff8786502be4d5158fd2a28b8b66fdf55700debaa4e67b074c', 'module': 'item_ids.py', 'body_sha256': 'sha256:2158491c6a50bfdca082dff40637c578a45e31fbf1c44c3d3847f8c88e747de3'}